refactor: split exalted utilities into modules
