+++ components/ExaltedCharacterManager.tsx
@@ -43,12 +43,8 @@ import type { Character, AttributeType, AbilityType, ArmorPiece } from "@/lib/ch
 import { createNewCharacter } from "@/lib/character-defaults"
 import { toast } from "sonner"
 import { v4 as uuidv4 } from "uuid"
-import {
-  getAnimaLevel,
-  getActiveAnimaRulings,
-  calculateStatTotal,
-  clampModifier,
-} from "@/lib/exalted-utils"
+import { getAnimaLevel, getActiveAnimaRulings } from "@/lib/anima";
+import { calculateStatTotal } from "@/lib/staticValues"
 import { RulingsTab } from "@/components/character-tabs/RulingsTab"
 import { PowersTab } from "@/components/character-tabs/PowersTab"
 import { SocialTab } from "@/components/character-tabs/SocialTab"
