+++ components/character-tabs/CombatTab.tsx
@@ -17,7 +17,8 @@ import { Badge } from "@/components/ui/badge"
 import { Separator } from "@/components/ui/separator"
 import { Textarea } from "@/components/ui/textarea"
 import type { Character, ExaltType, DramaticInjury } from "@/lib/character-types"
-import { getAnimaLevel, getActiveAnimaRulings, calculateStatTotal } from "@/lib/exalted-utils"
+import { getAnimaLevel, getActiveAnimaRulings } from "@/lib/anima"
+import { calculateStatTotal } from "@/lib/staticValues"
 import type { CharacterCalculations } from "@/hooks/useCharacterCalculations"
 import { v4 as uuidv4 } from "uuid"
